backend/requirements.txt
fastapi
uvicorn
python-dotenv
gigachat
psycopg
sentence-transformers
pymorphy3
natasha
keybert
